v1.4.1?

+++ src/Utils.hpp
@@ -1,9 +1,11 @@
 #pragma once
+#include <string>
+#include <Geode/cocos/base_nodes/CCNode.h>
 
 class Utils {
   public:
 	static int randomIndex(int size);
-	static int stoi(std::string);
+	static uint64_t stoi(std::string);
 	static bool isSupportedExtension(std::string);
 	static bool getBool(std::string);
 	static void removeCard();
